requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-api-python-client==2.160.0
google-auth==2.38.0
schedule==1.2.0
pytest==8.0.0
pytest-cov==4.1.0
